+++ Bibioteca/src/main/java/compass/microservice/biblioteca/controller/LivroController.java
@@ -14,6 +14,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,6 +31,7 @@ import compass.microservice.biblioteca.controller.form.AtualizarLivroForm;
 import compass.microservice.biblioteca.controller.form.CadastrarLivroForm;
 import compass.microservice.biblioteca.modelos.Biblioteca;
 import compass.microservice.biblioteca.modelos.Livro;
+import compass.microservice.biblioteca.modelos.StatusLivro;
 import compass.microservice.biblioteca.repository.BibliotecaRepository;
 import compass.microservice.biblioteca.repository.LivrosRepository;
 
@@ -40,11 +42,11 @@ public class LivroController {
 
 	@Autowired
 	private LivrosRepository lRepo;
-	
+
 	@Autowired
 	private BibliotecaRepository bRepo;
-	
-	
+
+
 	@GetMapping
 	@Cacheable(value = "listaDeLivros")
 	public Page<LivroDto> lista(@PageableDefault(sort="id", direction= Direction.ASC,page=0, size=10)Pageable paginacao) {
@@ -63,19 +65,20 @@ public class LivroController {
 			return ResponseEntity.notFound().build();
 		}
 	}
-	
+
 	@GetMapping("/biblioteca/{id}")
 	public Page<LivroDto> livrosBiblioteca(	@PathVariable Long id,
 			@PageableDefault(sort="id", direction= Direction.ASC,page=0, size=10)Pageable paginacao){
 
 		Page<Livro>  livros = lRepo.findByBiblioteca_id(id,paginacao);
-		
-		
+
+
 		return LivroDto.converter(livros);
 	}
-	
-	
+
+
 	@PostMapping("/cadastrar")
+	@PreAuthorize("hasRole('ROLE_BIBLIOTECA')")
 	@Transactional
 	@CacheEvict(value = "listaDeLivros", allEntries = true)
 	public ResponseEntity<?> cadastrarLivro(@RequestBody CadastrarLivroForm form , UriComponentsBuilder uriBuilder){
@@ -96,38 +99,45 @@ public class LivroController {
 		}
 
 	}
-	
+
 	@PutMapping("/{id}")
+	@PreAuthorize("hasRole('ROLE_BIBLIOTECA')")
 	@Transactional
 	public ResponseEntity<LivroDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizarLivroForm form) {	
-			
+
 		Optional<Livro> opt = lRepo.findById(id);
 
-			if (opt.isPresent()) {
-				Livro l = opt.get();
-				Livro livroAtualizado = form.atualizar(l, l.getBiblioteca());
-				return ResponseEntity.ok(new LivroDto (livroAtualizado));
-			}
-		
+		if (opt.isPresent()) {
+			Livro l = opt.get();
+			Livro livroAtualizado = form.atualizar(l, l.getBiblioteca());
+			return ResponseEntity.ok(new LivroDto (livroAtualizado));
+		}
+
 		return ResponseEntity.notFound().build();
 	}
 
 
 
 	@DeleteMapping("/{id}")
+	@PreAuthorize("hasRole('ROLE_BIBLIOTECA')")
 	@Transactional
 	@CacheEvict(value = "listaDeLivros", allEntries = true)
 	public ResponseEntity<?> remover(@PathVariable Long id) {
 		Optional<Livro> optional = lRepo.findById(id);
 		if (optional.isPresent()) {
-			lRepo.deleteById(id);
-			return ResponseEntity.ok().build();
+			if(optional.get().getStatusLivro().equals(StatusLivro.INDISPONIVEL)) {
+				return ResponseEntity.badRequest().body("Não é possível deletar livros que estejam indisponíveis");
+			}else {
+				lRepo.deleteById(id);
+				return ResponseEntity.ok().build();
+			}
+
 		} else {
 			return ResponseEntity.notFound().build();
 		}
 	}
-	
-	
-	
+
+
+
 
 }
